move lit styles to separate files
wip
