+++ source/_include/_layout.pug
@@ -18,7 +18,7 @@ html.no-js(lang="en")
     link(rel="manifest" href="site.webmanifest")
     link(rel="apple-touch-icon" href="icon.png")
     // Place favicon.ico in the root directory
-    link(rel="stylesheet" href="css/main.sass")
+    link(rel="stylesheet" href="./css/main.sass")
 
   body(
     data-page=page
@@ -34,7 +34,7 @@ html.no-js(lang="en")
     header#Header
       h1
         a(href="./")
-          img(src="/images/logotype.svg" alt="TogoDX - Togo Data Explorer")
+          img(src="./images/logotype.svg" alt="TogoDX - Togo Data Explorer")
       nav
         ul
           li
