+++ source/css/object/component/_loading-view.sass
@@ -1,36 +1,3 @@
+@use "../../foundation/_mixins.sass" as *
 .loading-view
-  position: absolute
-  width: 100%
-  height: 100%
-  top: 0
-  left: 0
-  background-color: transparent
-  opacity: 0
-  pointer-events: none
-  transition: background var(--transition-wipe), opacity var(--transition-wipe)
-  &::before
-    content: "autorenew"
-    font-size: 1.2rem
-    font-family: "Material Icons Outlined"
-    text-align: center
-    display: block
-    width: 1.6rem
-    height: 1.6rem
-    line-height: 1.6rem
-    border-radius: .8rem
-    background-color: rgba(white, .8)
-    position: absolute
-    left: calc(50% - .8rem)
-    top: calc(50% - .8rem)
-  &.-shown
-    background-color: var(--color-translucent-lamp-black3)
-    opacity: 1
-    pointer-events: auto
-    &::before
-      animation: loading 2s linear infinite
-
-@keyframes loading
-  from
-    transform: rotate(0deg)
-  to
-    transform: rotate(360deg)
+  +loading-view()
