+++ source/js/classes/AttributeTrackView.js
@@ -6,7 +6,6 @@ import ColumnSelectorView from './ColumnSelectorView';
 import HistogramRangeSelectorView from './HistogramRangeSelectorView';
 import TrackOverviewCategorical from './TrackOverviewCategorical';
 import ConditionAnnotation from './ConditionAnnotation';
-import {Dummy} from './CategoryBrowser/dummy';
 import * as event from '../events';
 import {CategoryBrowser} from './CategoryBrowser/CategoryBrowser';
 
@@ -208,12 +207,15 @@ export default class AttributeTrackView {
     // make selector view
     switch (this.#attribute.datamodel) {
       case 'classification':
-        // new ColumnSelectorView(
-        //   this.#SELECT_CONTAINER,
-        //   this.#attribute,
-        //   filters
-        // );
-        new CategoryBrowser(this.#SELECT_CONTAINER, this.#attribute, filters);
+        if (filters.some(filter => !filter.tip)) {
+          new CategoryBrowser(this.#SELECT_CONTAINER, this.#attribute, filters);
+        } else {
+          new ColumnSelectorView(
+            this.#SELECT_CONTAINER,
+            this.#attribute,
+            filters
+          );
+        }
         break;
       case 'distribution':
         new HistogramRangeSelectorView(
