--- source/js/classes/CategoryBrowser/dummy.js
@@ -1,16 +1,0 @@
-export class Dummy {
-  #elm;
-  #attribute;
-  #items;
-
-  constructor(elm, attribute, items) {
-    this.#elm = elm;
-    this.#attribute = attribute;
-    this.#items = items;
-    console.log({elm, attribute, items});
-  }
-
-  get #api() {
-    return this.#elm.dataset.api + '?hierarchy';
-  }
-}
